Fix settings modals: inline tokens, real org data, drop overview chevron
- Replace the "Open token manager" redirect with an inline TokensManager
  in the Account Settings modal so users no longer get bounced to a
  full-page route that looked sparse on wide screens.
- Refactor TokensPage to reuse the same TokensManager component, wrapped
  in a centered card so the legacy /account/tokens URL still renders well.
- Rewrite the Organization Settings modal to show real tenant data
  (display name, slug, role, store count) sourced from StoreContext,
  and replace the mock "MCP Servers" tab with a real "Stores" list
  filtered to the active tenant. Editing is read-only since tenant
  updates require global admin via /admin/tenants.
- Remove the decorative ChevronRight from the file tree's Overview
  button.
