+++ stash_ui/src/app/components/OrganizationSettingsModal.tsx
@@ -1,34 +1,30 @@
 import React, { useState } from "react";
-import {
-  X,
-  Server,
-  Building2,
-  Trash2,
-  FolderTree,
-  Lock,
-  LockOpen,
-  Edit2,
-} from "lucide-react";
-import {
-  CreateServerModal,
-  ServerConfig,
-} from "./CreateServerModal";
+import { X, Building2, Database, Shield } from "lucide-react";
+import { useStore } from "../StoreContext";
+import { StoreSummary } from "../../api/auth";
 
 interface OrganizationSettingsModalProps {
   isOpen: boolean;
   onClose: () => void;
 }
 
-type SettingsTab = "general" | "servers";
+type SettingsTab = "general" | "stores";
 
 export function OrganizationSettingsModal({
   isOpen,
   onClose,
 }: OrganizationSettingsModalProps) {
-  const [activeTab, setActiveTab] =
-    useState<SettingsTab>("general");
+  const [activeTab, setActiveTab] = useState<SettingsTab>("general");
+  const { current, stores } = useStore();
 
   if (!isOpen) return null;
+
+  // Show only stores in the currently-selected tenant. If there's no
+  // active store the modal can't infer which org you mean, so render an
+  // empty-state instead of leaking other tenants' data.
+  const tenantStores = current
+    ? stores.filter((s) => s.tenant_slug === current.tenant_slug)
+    : [];
 
   const tabs = [
     {
@@ -37,9 +33,9 @@ export function OrganizationSettingsModal({
       icon: Building2,
     },
     {
-      id: "servers" as SettingsTab,
-      label: "MCP Servers",
-      icon: Server,
+      id: "stores" as SettingsTab,
+      label: "Stores",
+      icon: Database,
     },
   ];
 
@@ -50,14 +46,13 @@ export function OrganizationSettingsModal({
       onClick={onClose}
     >
       <div
-        className="w-full max-w-5xl h-[85vh] flex flex-col rounded-lg shadow-2xl overflow-hidden"
+        className="w-full max-w-4xl h-[80vh] flex flex-col rounded-lg shadow-2xl overflow-hidden"
         style={{
           backgroundColor: "var(--stash-bg-surface)",
           border: "1px solid var(--stash-border)",
         }}
         onClick={(e) => e.stopPropagation()}
       >
-        {/* Header */}
         <div
           className="flex items-center justify-between px-6 py-4 border-b"
           style={{ borderColor: "var(--stash-border)" }}
@@ -73,21 +68,17 @@ export function OrganizationSettingsModal({
             className="p-2 rounded transition-all duration-150"
             style={{ color: "var(--stash-text-secondary)" }}
             onMouseEnter={(e) => {
-              e.currentTarget.style.backgroundColor =
-                "var(--stash-bg-hover)";
+              e.currentTarget.style.backgroundColor = "var(--stash-bg-hover)";
             }}
             onMouseLeave={(e) => {
-              e.currentTarget.style.backgroundColor =
-                "transparent";
+              e.currentTarget.style.backgroundColor = "transparent";
             }}
           >
             <X className="w-5 h-5" />
           </button>
         </div>
 
-        {/* Content with Sidebar Navigation */}
         <div className="flex flex-1 overflow-hidden">
-          {/* Sidebar Navigation */}
           <div
             className="w-56 border-r overflow-y-auto"
             style={{
@@ -120,8 +111,7 @@ export function OrganizationSettingsModal({
                     }}
                     onMouseLeave={(e) => {
                       if (!isActive) {
-                        e.currentTarget.style.backgroundColor =
-                          "transparent";
+                        e.currentTarget.style.backgroundColor = "transparent";
                       }
                     }}
                   >
@@ -133,10 +123,13 @@ export function OrganizationSettingsModal({
             </nav>
           </div>
 
-          {/* Main Content Area */}
           <div className="flex-1 overflow-y-auto p-6">
-            {activeTab === "general" && <GeneralSettings />}
-            {activeTab === "servers" && <ServerSettings />}
+            {activeTab === "general" && (
+              <GeneralSettings current={current} storeCount={tenantStores.length} />
+            )}
+            {activeTab === "stores" && (
+              <StoresList current={current} stores={tenantStores} />
+            )}
           </div>
         </div>
       </div>
@@ -144,428 +137,208 @@ export function OrganizationSettingsModal({
   );
 }
 
-function GeneralSettings() {
+function GeneralSettings({
+  current,
+  storeCount,
+}: {
+  current: StoreSummary | null;
+  storeCount: number;
+}) {
+  if (!current) {
+    return (
+      <EmptyOrgState message="Select a store to view its organization." />
+    );
+  }
   return (
     <div>
       <h3
-        className="text-base font-semibold mb-4"
+        className="text-base font-semibold mb-1"
         style={{ color: "var(--stash-text-bright)" }}
       >
-        General Settings
+        General
       </h3>
+      <p
+        className="text-sm mb-6"
+        style={{ color: "var(--stash-text-secondary)" }}
+      >
+        Tenant metadata is managed by global admins via the{" "}
+        <code>/admin/tenants</code> API. The fields below are read-only here.
+      </p>
       <div className="space-y-4">
+        <ReadOnlyField label="Display name" value={current.tenant_display_name} />
+        <ReadOnlyField label="Slug" value={current.tenant_slug} mono />
+        <ReadOnlyField
+          label="Stores in this organization"
+          value={String(storeCount)}
+        />
         <div>
           <label
-            className="block text-sm mb-2"
+            className="block text-sm mb-2 flex items-center gap-2"
             style={{ color: "var(--stash-text-primary)" }}
           >
-            Organization Name
+            <Shield className="w-4 h-4" />
+            Your role
           </label>
-          <input
-            type="text"
-            defaultValue="My Organization"
-            className="w-full px-3 py-2 rounded-md text-sm outline-none transition-all duration-150"
+          <span
+            className="inline-block text-xs px-2 py-1 rounded uppercase tracking-wide"
             style={{
               backgroundColor: "var(--stash-bg-base)",
-              color: "var(--stash-text-primary)",
-              border: "1px solid var(--stash-border)",
+              color:
+                current.role === "admin"
+                  ? "var(--stash-accent)"
+                  : "var(--stash-text-secondary)",
+              border: `1px solid ${
+                current.role === "admin"
+                  ? "var(--stash-accent)"
+                  : "var(--stash-border)"
+              }`,
             }}
-            onFocus={(e) => {
-              e.currentTarget.style.borderColor =
-                "var(--stash-accent)";
-              e.currentTarget.style.boxShadow =
-                "0 0 0 2px rgba(148, 226, 213, 0.1)";
-            }}
-            onBlur={(e) => {
-              e.currentTarget.style.borderColor =
-                "var(--stash-border)";
-              e.currentTarget.style.boxShadow = "none";
-            }}
-          />
-        </div>
-        <div>
-          <label
-            className="block text-sm mb-2"
-            style={{ color: "var(--stash-text-primary)" }}
           >
-            Description
-          </label>
-          <textarea
-            defaultValue="Manage your MCP servers and content roots"
-            rows={3}
-            className="w-full px-3 py-2 rounded-md text-sm outline-none transition-all duration-150 resize-none"
-            style={{
-              backgroundColor: "var(--stash-bg-base)",
-              color: "var(--stash-text-primary)",
-              border: "1px solid var(--stash-border)",
-            }}
-            onFocus={(e) => {
-              e.currentTarget.style.borderColor =
-                "var(--stash-accent)";
-              e.currentTarget.style.boxShadow =
-                "0 0 0 2px rgba(148, 226, 213, 0.1)";
-            }}
-            onBlur={(e) => {
-              e.currentTarget.style.borderColor =
-                "var(--stash-border)";
-              e.currentTarget.style.boxShadow = "none";
-            }}
-          />
+            {current.role}
+          </span>
         </div>
       </div>
     </div>
   );
 }
 
-function ServerSettings() {
-  const [servers, setServers] = useState<ServerConfig[]>([]);
-  const [isCreateModalOpen, setIsCreateModalOpen] =
-    useState(false);
-  const [editingServer, setEditingServer] =
-    useState<ServerConfig | null>(null);
-
-  const handleCreateServer = (server: ServerConfig) => {
-    setServers([...servers, server]);
-  };
-
-  const handleUpdateServer = (updatedServer: ServerConfig) => {
-    setServers(
-      servers.map((s) =>
-        s.id === updatedServer.id ? updatedServer : s,
-      ),
+function StoresList({
+  current,
+  stores,
+}: {
+  current: StoreSummary | null;
+  stores: StoreSummary[];
+}) {
+  if (!current) {
+    return (
+      <EmptyOrgState message="Select a store to view its organization." />
     );
-    setEditingServer(null);
-  };
-
-  const handleDeleteServer = (serverId: string) => {
-    setServers(servers.filter((s) => s.id !== serverId));
-  };
-
-  const handleEditServer = (server: ServerConfig) => {
-    setEditingServer(server);
-    setIsCreateModalOpen(true);
-  };
-
-  const handleCloseModal = () => {
-    setIsCreateModalOpen(false);
-    setEditingServer(null);
-  };
-
+  }
   return (
-    <>
-      <div>
-        <div className="flex items-center justify-between mb-4">
-          <h3
-            className="text-base font-semibold"
-            style={{ color: "var(--stash-text-bright)" }}
-          >
-            MCP Servers
-          </h3>
-          <button
-            onClick={() => setIsCreateModalOpen(true)}
-            className="px-4 py-2 rounded-md text-sm transition-all duration-150"
-            style={{
-              backgroundColor: "var(--stash-accent)",
-              color: "var(--stash-bg-base)",
-            }}
-            onMouseEnter={(e) => {
-              e.currentTarget.style.opacity = "0.9";
-            }}
-            onMouseLeave={(e) => {
-              e.currentTarget.style.opacity = "1";
-            }}
-          >
-            + Create Server
-          </button>
-        </div>
-        <p
-          className="text-sm mb-6"
-          style={{ color: "var(--stash-text-secondary)" }}
+    <div>
+      <h3
+        className="text-base font-semibold mb-1"
+        style={{ color: "var(--stash-text-bright)" }}
+      >
+        Stores
+      </h3>
+      <p
+        className="text-sm mb-4"
+        style={{ color: "var(--stash-text-secondary)" }}
+      >
+        Content stores in <strong>{current.tenant_display_name}</strong>.
+        Provisioning new stores is handled by global admins via the{" "}
+        <code>/admin/tenants/{"{id}"}/stores</code> API.
+      </p>
+      {stores.length === 0 ? (
+        <div
+          className="p-6 rounded-md text-center text-sm"
+          style={{
+            backgroundColor: "var(--stash-bg-base)",
+            border: "1px dashed var(--stash-border)",
+            color: "var(--stash-text-secondary)",
+          }}
         >
-          Manage MCP servers that can access your content roots
-        </p>
-
-        {servers.length === 0 ? (
-          <div
-            className="p-6 rounded-md text-center"
-            style={{
-              backgroundColor: "var(--stash-bg-base)",
-              border: "1px dashed var(--stash-border)",
-            }}
-          >
-            <Server
-              className="w-12 h-12 mx-auto mb-3"
-              style={{ color: "var(--stash-text-secondary)" }}
-            />
-            <p
-              className="text-sm"
-              style={{ color: "var(--stash-text-secondary)" }}
+          No stores visible to you in this organization.
+        </div>
+      ) : (
+        <div
+          className="rounded-md overflow-hidden"
+          style={{ border: "1px solid var(--stash-border)" }}
+        >
+          {stores.map((s, idx) => (
+            <div
+              key={s.id}
+              className="flex items-center justify-between px-4 py-3 text-sm"
+              style={{
+                backgroundColor: "var(--stash-bg-base)",
+                borderTop: idx === 0 ? "none" : "1px solid var(--stash-border)",
+              }}
             >
-              No MCP servers configured yet
-            </p>
-          </div>
-        ) : (
-          <div className="space-y-3">
-            {servers.map((server) => (
-              <div
-                key={server.id}
-                className="p-4 rounded-md border"
+              <div className="min-w-0">
+                <div
+                  className="font-medium"
+                  style={{ color: "var(--stash-text-bright)" }}
+                >
+                  {s.display_name}
+                </div>
+                <code
+                  className="text-xs"
+                  style={{ color: "var(--stash-text-secondary)" }}
+                >
+                  /{s.tenant_slug}/{s.slug}
+                </code>
+              </div>
+              <span
+                className="text-xs px-2 py-0.5 rounded uppercase tracking-wide"
                 style={{
-                  backgroundColor: "var(--stash-bg-base)",
-                  borderColor: "var(--stash-border)",
+                  backgroundColor: "var(--stash-bg-surface)",
+                  color:
+                    s.role === "admin"
+                      ? "var(--stash-accent)"
+                      : "var(--stash-text-secondary)",
+                  border: `1px solid ${
+                    s.role === "admin"
+                      ? "var(--stash-accent)"
+                      : "var(--stash-border)"
+                  }`,
                 }}
               >
-                <div className="flex items-start justify-between mb-2">
-                  <div className="flex-1">
-                    <div className="flex items-center gap-2 mb-1">
-                      <h4
-                        className="font-medium"
-                        style={{
-                          color: "var(--stash-text-bright)",
-                        }}
-                      >
-                        {server.name}
-                      </h4>
-                    </div>
-                    {server.description && (
-                      <p
-                        className="text-sm mb-2"
-                        style={{
-                          color: "var(--stash-text-secondary)",
-                        }}
-                      >
-                        {server.description}
-                      </p>
-                    )}
-                    <div
-                      className="text-xs mb-3"
-                      style={{
-                        color: "var(--stash-text-secondary)",
-                      }}
-                    >
-                      Timeout: {server.timeout}s
-                    </div>
+                {s.role}
+              </span>
+            </div>
+          ))}
+        </div>
+      )}
+    </div>
+  );
+}
 
-                    {/* Content Roots */}
-                    {server.contentRoots &&
-                      server.contentRoots.length > 0 && (
-                        <div className="mt-3">
-                          <div className="flex items-center gap-2 mb-2">
-                            <FolderTree
-                              className="w-3.5 h-3.5"
-                              style={{
-                                color: "var(--stash-accent)",
-                              }}
-                            />
-                            <span
-                              className="text-xs font-medium"
-                              style={{
-                                color:
-                                  "var(--stash-text-primary)",
-                              }}
-                            >
-                              Content Root
-                            </span>
-                          </div>
-                          <div className="space-y-2">
-                            {server.contentRoots.map(
-                              (contentRoot) => (
-                                <div
-                                  key={contentRoot.id}
-                                  className="p-2 rounded-md"
-                                  style={{
-                                    backgroundColor:
-                                      "var(--stash-bg-surface)",
-                                    border:
-                                      "1px solid var(--stash-border)",
-                                  }}
-                                >
-                                  <div className="flex items-center gap-2 mb-1">
-                                    <span
-                                      className="text-xs font-medium"
-                                      style={{
-                                        color:
-                                          "var(--stash-text-bright)",
-                                      }}
-                                    >
-                                      {contentRoot.name ||
-                                        "Unnamed Content Root"}
-                                    </span>
-                                    <span
-                                      className="text-xs px-1.5 py-0.5 rounded"
-                                      style={{
-                                        backgroundColor:
-                                          "var(--stash-bg-base)",
-                                        color:
-                                          "var(--stash-accent)",
-                                        border:
-                                          "1px solid var(--stash-accent)",
-                                      }}
-                                    >
-                                      {contentRoot.type ===
-                                      "simple"
-                                        ? "Simple"
-                                        : "Virtual"}
-                                    </span>
-                                  </div>
-                                  {contentRoot.description && (
-                                    <p
-                                      className="text-xs mb-1"
-                                      style={{
-                                        color:
-                                          "var(--stash-text-secondary)",
-                                      }}
-                                    >
-                                      {contentRoot.description}
-                                    </p>
-                                  )}
+function EmptyOrgState({ message }: { message: string }) {
+  return (
+    <div
+      className="p-6 rounded-md text-sm"
+      style={{
+        backgroundColor: "var(--stash-bg-base)",
+        border: "1px dashed var(--stash-border)",
+        color: "var(--stash-text-secondary)",
+      }}
+    >
+      {message}
+    </div>
+  );
+}
 
-                                  {/* Simple Directory */}
-                                  {contentRoot.type ===
-                                    "simple" &&
-                                    contentRoot.path && (
-                                      <code
-                                        className="text-xs px-1 py-0.5 rounded"
-                                        style={{
-                                          backgroundColor:
-                                            "var(--stash-bg-base)",
-                                          color:
-                                            "var(--stash-text-primary)",
-                                        }}
-                                      >
-                                        {contentRoot.path}
-                                      </code>
-                                    )}
-
-                                  {/* Virtual Content Root */}
-                                  {contentRoot.type ===
-                                    "virtual" &&
-                                    contentRoot.mountedDirectories &&
-                                    contentRoot
-                                      .mountedDirectories
-                                      .length > 0 && (
-                                      <div className="mt-1 space-y-1">
-                                        {contentRoot.mountedDirectories.map(
-                                          (mount) => (
-                                            <div
-                                              key={mount.id}
-                                              className="flex items-center justify-between text-xs p-1.5 rounded"
-                                              style={{
-                                                backgroundColor:
-                                                  "var(--stash-bg-base)",
-                                              }}
-                                            >
-                                              <code
-                                                style={{
-                                                  color:
-                                                    "var(--stash-text-primary)",
-                                                }}
-                                              >
-                                                {mount.path}
-                                              </code>
-                                              <div className="flex items-center gap-1">
-                                                {mount.permission ===
-                                                "read" ? (
-                                                  <Lock
-                                                    className="w-3 h-3"
-                                                    style={{
-                                                      color:
-                                                        "var(--stash-text-secondary)",
-                                                    }}
-                                                  />
-                                                ) : (
-                                                  <LockOpen
-                                                    className="w-3 h-3"
-                                                    style={{
-                                                      color:
-                                                        "var(--stash-accent)",
-                                                    }}
-                                                  />
-                                                )}
-                                                <span
-                                                  style={{
-                                                    color:
-                                                      "var(--stash-text-secondary)",
-                                                  }}
-                                                >
-                                                  {mount.permission ===
-                                                  "read"
-                                                    ? "Read"
-                                                    : "Read-Write"}
-                                                </span>
-                                              </div>
-                                            </div>
-                                          ),
-                                        )}
-                                      </div>
-                                    )}
-                                </div>
-                              ),
-                            )}
-                          </div>
-                        </div>
-                      )}
-                  </div>
-                  <div className="flex items-center gap-2">
-                    <button
-                      className="p-2 rounded transition-all duration-150"
-                      style={{
-                        color: "var(--stash-text-secondary)",
-                      }}
-                      onMouseEnter={(e) => {
-                        e.currentTarget.style.backgroundColor =
-                          "var(--stash-bg-hover)";
-                      }}
-                      onMouseLeave={(e) => {
-                        e.currentTarget.style.backgroundColor =
-                          "transparent";
-                      }}
-                      title="Edit server"
-                      onClick={() => handleEditServer(server)}
-                    >
-                      <Edit2
-                        className="w-4 h-4"
-                        style={{ color: "var(--stash-accent)" }}
-                      />
-                    </button>
-                    <button
-                      className="p-2 rounded transition-all duration-150"
-                      style={{
-                        color: "var(--stash-text-secondary)",
-                      }}
-                      onMouseEnter={(e) => {
-                        e.currentTarget.style.backgroundColor =
-                          "var(--stash-bg-hover)";
-                      }}
-                      onMouseLeave={(e) => {
-                        e.currentTarget.style.backgroundColor =
-                          "transparent";
-                      }}
-                      title="Delete server"
-                      onClick={() =>
-                        handleDeleteServer(server.id)
-                      }
-                    >
-                      <Trash2
-                        className="w-4 h-4"
-                        style={{ color: "var(--stash-error)" }}
-                      />
-                    </button>
-                  </div>
-                </div>
-              </div>
-            ))}
-          </div>
-        )}
-      </div>
-
-      <CreateServerModal
-        isOpen={isCreateModalOpen}
-        onClose={handleCloseModal}
-        onCreate={handleCreateServer}
-        onUpdate={handleUpdateServer}
-        editingServer={editingServer}
+function ReadOnlyField({
+  label,
+  value,
+  mono = false,
+}: {
+  label: string;
+  value: string;
+  mono?: boolean;
+}) {
+  return (
+    <div>
+      <label
+        className="block text-sm mb-2"
+        style={{ color: "var(--stash-text-primary)" }}
+      >
+        {label}
+      </label>
+      <input
+        type="text"
+        value={value}
+        readOnly
+        className={`w-full px-3 py-2 rounded-md text-sm outline-none ${
+          mono ? "font-mono" : ""
+        }`}
+        style={{
+          backgroundColor: "var(--stash-bg-base)",
+          color: "var(--stash-text-primary)",
+          border: "1px solid var(--stash-border)",
+        }}
       />
-    </>
+    </div>
   );
 }
